ref(wallet): move getting local utxos to `create_tx`
We can't easily plan the utxos before all of the assets are known,
so having `TxBuilder` try to create `WeightedUtxos` creates a
chicken and egg problem that we fix by moving the logic to
`Wallet::create_tx`.

`TxParams::utxos` field is thus split into two: `utxos` is a list
of outpoints strictly to be used for getting locally owned outputs,
and `foreign_utxos` handles adding `WeightedUtxos` coming from
`add_foreign_utxo`. These are later combined with the manually
selected coins to form the list of utxos that must be spent.

The other side effect of this change is that `build_fee_bump` now
adds all of the original inputs back to the tx builder as
foreign utxos since it knows all of the the satisfaction weights.