fix(esplora): `chain_update` errors if no point of connection
Before, the `chain_update` function might have panicked if the
local checkpoint was not on the same network as the remote
server. Now if we have iterated all of the blocks of the local
CP and do not find a "point of agreement", then we return early
with a `esplora_client::Error::HeaderHashNotFound`.

- Also refactor to avoid the use of .expect during `fetch_block`
function.

- Also add a test to cover the expected code paths.