fix(keychain_txout): Remove `spk_cache_stage`
`spk_cache_stage` was being used to hold newly derived
SPKs before being transferred to the `ChangeSet`. As a result
it became necessary to call `_empty_stage_into_changeset` whenever
the internal state changed, requiring the developer to know
which methods depended on the functionality, e.g. `scan_txout`
and `reveal_next_spk`.

Further, this altered the behavior of the `ChangeSet` in
surprising ways. For example revealing 1 SPK might return
additional scripts in the `ChangeSet` if they had
previously been staged.

The rationale in #1963 was to avoid returning a `ChangeSet` from
`insert_descriptor`. We can do this by providing another
method that returns a tuple of `(bool, ChangeSet)` so that
descriptor insertion results in updating the `ChangeSet` directly.
Any code using the original `insert_descriptor` can still access the
`ChangeSet` using `initial_changeset` (which may include SPKs
from other descriptors), or we can expose an accessor to get a
reference to the `spk_cache`.